Moved BuildMessageService to a more fitting location
